Fix for bug where context changes were not merged.
When `NSInMemoryStoreType` is used and a private context is saved, merging
the changes into the main context hangs. By wrapping the merge call in a
`performBlock:` on the private context being saved solves this problem.

A failing unit test (subsequent to the described change) is removed. This
is due to the `mergeChangesFromContextDidSaveNotification:` method not
being called immediately anymore; is called from inside the block.
Getting it to pass will mean involve suspending the run loop while the
`performBlock:` call has finished. Maybe use something like this:
https://github.com/hfossli/AGAsyncTestHelper

Former-commit-id: 3e8485ac940ce980ec5c3ea57562e9e4ed469113