Changes merged only when a private context is saved.
When merging changes from a private context into the main context,
the `mergeChangesFromContextDidSaveNotification` call is done from
within the `performBlock:` on the private context. However, if
`performBlock:` is used on the main context then the following
exception is thrown:
"Can only use -performBlock: on an NSManagedObjectContext that was
created with a queue."
So when the `NSManagedObjectContextDidSaveNotification` notification
if received, the context being saved is checked to see if it the main
context. If it is, then it is not necessary to merge any changes. If it
isn't then the changes must be merged from the private context
(passed with the notification) to the main context.


Former-commit-id: 8aa5e4605033b8a71baab9464661487c376612d4