Fix for merging changes when main context is saved.
Apparently testing pointer equality for NSManagedObjectContexts
doesn't work when checking if the NSManagedObjectContexts in the
notification is same as the stored context stored in the
`mainContext` property of SQKContextManager. (Maybe some weird
threading thing as NSOperations are being used? I wouldn't be at
all surprised.)
Workaround is to check the concurrency type. Private contexts are
always NSPrivateQueueConcurrencyType, so in these case we need to
merge.


Former-commit-id: 7ddff7fbde452a7ae36e64f4f9309f12377f2a21