Renamed variable from `sut`, i.e. the System Under Test.
Although more concise sut is not clear enough as a variable name.
A comment explaining the acronym was placed at the property declaration.
However upon further consideration it it thought that if a shortened variable
name needs a comment to explain it, then what is the point in shortening it
in the first place?