sqk_insertOrUpdate: supports inserting stub objects when only primary key values are known.
It is often the case that you only know the GUIDs of objects when
working with data from a web service. Say for example a user has a number
of commits, but the JSON object for the user only specifies an array of GUIDs
of those commits rather than the full commit objects themselves. e.g.:

{
	"userID" : "9c9c4eab-7665-4b8e-b4cc-9d1f9af27926",
	"userName" : "Luke Stringer",
	"commitIDs" : [
		"c800a783-38da-4639-80dd-71a8baee1a31",
		"da0cd9ee-ef78-48ee-b60a-727c615ddd98",
		"9fbe8b1c-2149-4317-9d46-92b8ec65b0ae",
	]
}

In this case you would probably want to insert-or-update 'stub' commit objects that
only have the GUID set, and then request the data from a /commit endpoint
at a later date. This update to the `sqk_insertOrUpdate:` method facilities
this by passing @"self" as the `uniqueRemoteKey:` parameter, that would usually
be used to pull out a value from a dictionary.