Fixed crash with SQKAlternateCommitsViewController
The crash was caused by the context changing but the SQKAlternateCommitsViewController's implementation of NSFetchedResultsController not having performed the fetch so it didn't know of the objects that had been inserted / updated.
The fetch was only done when the view is loaded. The view is only loaded when first navigated to the tab.

The fix was to only set the delegate of the controller once the perfomFetch can be done.